+++ tests/sgmcmc/test_sghmc.py
@@ -2,7 +2,8 @@ from functools import partial
 import torch
 from posteriors.sgmcmc import sghmc
 from tests import scenarios
-from tests.sgmcmc.utils import run_test_mcmc_gaussian
+from tests.utils import verify_inplace_update
+from tests.sgmcmc.utils import run_test_sgmcmc_gaussian
 
 
 def test_sghmc():
@@ -10,17 +11,21 @@ def test_sghmc():
 
     # Set inference parameters
     lr = 1e-2
-    alpha = 0.01
+    alpha = 0.1
     sigma = 1.0
     temperature = 1.0
-    n_steps = 10_000
-    burnin = 1_000
+    beta = 0.0
 
     # Run MCMC test on Gaussian
-    run_test_mcmc_gaussian(
-        partial(sghmc.build, lr=lr, alpha=alpha, sigma=sigma, temperature=temperature),
-        n_steps=n_steps,
-        burnin=burnin,
+    run_test_sgmcmc_gaussian(
+        partial(
+            sghmc.build,
+            lr=lr,
+            alpha=alpha,
+            sigma=sigma,
+            temperature=temperature,
+            beta=beta,
+        ),
     )
 
 
@@ -41,4 +46,4 @@ def test_sghmc_inplace_step():
     params = torch.randn(dim)
 
     # Verify inplace update
-    scenarios.verify_inplace_update(transform, params, None)
+    verify_inplace_update(transform, params, None)
