+++ tests/sgmcmc/test_sgld.py
@@ -2,7 +2,7 @@ from functools import partial
 import torch
 from posteriors.sgmcmc import sgld
 from tests import scenarios
-from tests.sgmcmc.utils import run_test_mcmc_gaussian
+from tests.sgmcmc.utils import run_test_sgmcmc_gaussian
 
 
 def test_sgld():
@@ -10,17 +10,12 @@ def test_sgld():
 
     # Set inference parameters
     lr = 1e-2
-    alpha = 0.01
-    sigma = 1.0
+    beta = 0.0
     temperature = 1.0
-    n_steps = 10_000
-    burnin = 1_000
 
     # Run MCMC test on Gaussian
-    run_test_mcmc_gaussian(
-        partial(sgld.build, lr=lr, alpha=alpha, sigma=sigma, temperature=temperature),
-        n_steps=n_steps,
-        burnin=burnin,
+    run_test_sgmcmc_gaussian(
+        partial(sgld.build, lr=lr, beta=beta, temperature=temperature),
     )
